Rename report -> analytics
