+++ beemo/cli.py
@@ -54,8 +54,8 @@ def do_logs(
     run(input, csv_dir, pattern)
 
 
-@app.command("report")
-def do_report(
+@app.command("analytics")
+def do_analytics(
     csv_dir: Annotated[Optional[Path], typer.Option(help="Input CSV directory")] = None,
     templates_dir: Annotated[
         Optional[Path], typer.Option(help="Chameleon templates directory")
@@ -65,26 +65,29 @@ def do_report(
     base_url: Annotated[
         Optional[str], typer.Option(help="Site base URL e.g. https://bennuttall.com")
     ] = None,
-    title: Annotated[Optional[str], typer.Option(help="Report title")] = None,
+    title: Annotated[Optional[str], typer.Option(help="Analytics title")] = None,
 ):
-    """Generate HTML analytics report from log CSVs."""
-    from .reports.cli import run
+    """Generate HTML analytics site from log CSVs."""
+    from .analytics.cli import run
 
     config = _get_config()
-    report_config = config.report if config else None
+    analytics_config = config.analytics if config else None
     build_config = config.build if config else None
 
     csv_dir = _require(
-        csv_dir or (report_config.csv_dir if report_config else None), "report.csv_dir"
+        csv_dir or (analytics_config.csv_dir if analytics_config else None), "analytics.csv_dir"
     )
     templates_dir = _require(
         templates_dir or (build_config.templates_dir if build_config else None),
         "build.templates_dir",
     )
     manifest = manifest or (build_config.output_dir / "manifest.json" if build_config else None)
-    output_dir = _require(output_dir or (report_config.output_dir if report_config else None), "report.output_dir")
-    base_url = base_url or (report_config.base_url if report_config else "")
-    title = title or (report_config.title if report_config else "")
+    output_dir = _require(
+        output_dir or (analytics_config.output_dir if analytics_config else None),
+        "analytics.output_dir",
+    )
+    base_url = base_url or (analytics_config.base_url if analytics_config else "")
+    title = title or (analytics_config.title if analytics_config else "")
 
     run(csv_dir, templates_dir, manifest, output_dir, base_url, title)
